Move errors to module level.
