+++ verto/errors/StyleError.py
@@ -1,13 +1,13 @@
-from verto.processors.errors.Error import Error
+from verto.errors.Error import Error
 
 
 class StyleError(Error):
     '''Exception raised when a Style rule is broken.
 
     Attributes:
-        rule -- rule which was broken
-        lines -- lines where the style rule was broken
-        message -- explanation of why error was thrown
+        line_nums: the line numbers the rule as broken on
+        lines: lines where the style rule was broken
+        message: explanation of why error was thrown
     '''
 
     def __init__(self, line_nums, lines, message):
