+++ verto/errors/__init__.py
