+++ verto/processors/ConditionalProcessor.py
@@ -1,5 +1,5 @@
 from markdown.blockprocessors import BlockProcessor
-from verto.processors.errors.TagNotMatchedError import TagNotMatchedError
+from verto.errors.TagNotMatchedError import TagNotMatchedError
 from verto.processors.utils import etree, parse_arguments, parse_flag, blocks_to_string
 from verto.utils.HtmlParser import HtmlParser
 from collections import OrderedDict
